* uint-phase:
  TODO--
  TESTS: testifftsynth: bump number of saw_perf runs
  TESTS: testifftsynth: port to IFFTSynth uint phase API
  LIB: LiveDecoder: use uint for phase instead of float (performance)
  LIB: IFFTSynth: use uint for phase instead of float (performance)

+++ lib/smlivedecoder.hh
@@ -28,7 +28,7 @@ class LiveDecoder
   {
     float freq;
     float mag;
-    float phase;
+    uint  phase;
   };
   std::vector<PartialState> pstate[2], *last_pstate;
 
@@ -78,7 +78,7 @@ class LiveDecoder
 
   // unison
   int                 unison_voices;
-  std::vector<float>  unison_phases[2];
+  std::vector<uint>   unison_phases[2];
   std::vector<float>  unison_freq_factor;
   float               unison_gain;
